fix(node-core): read the correct vergen sha var and add the agent string
The version module read `VERTEX_GIT_SHA`, but build.rs emits `VERGEN_GIT_SHA`,
so `GIT_SHA` was always "unknown" and the derived long version was dead. Read
the var build.rs actually emits, drop the genuinely-unused consts (`USER_AGENT`,
`P2P_CLIENT_VERSION`, `SHORT_VERSION`, `BUILD_TIMESTAMP`, `CARGO_FEATURES`), and
add `AGENT_VERSION` (`vertex/<version>-<sha>`, bare `vertex/<version>` for a
git-less build) as the single libp2p agent string source. build.rs now emits
only the sha, with the Docker-safe fallback.
